--- src/assets/images/icon/HLogo.js
@@ -1,32 +1,0 @@
-import React from "react";
-export const WebLogo = () => {
-  return (
-    <svg
-      width="64"
-      height="64"
-      viewBox="0 0 64 64"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M0 32C0 14.3269 14.328 0 32 0C49.6731 0 64 14.328 64 32C64 49.6731 49.672 64 32 64C14.3269 64 0 49.672 0 32ZM3.55556 32C3.55556 16.2906 16.2871 3.55556 32 3.55556C47.7094 3.55556 60.4444 16.2871 60.4444 32C60.4444 47.7094 47.7129 60.4444 32 60.4444C16.2906 60.4444 3.55556 47.7129 3.55556 32Z"
-        fill="#1C6DD0"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M42.5377 41.7643V21.4622H37.8542V29.6104H26.1456V21.4622H21.4622V38.7083C22.9045 38.6636 24.4625 38.701 26.1456 38.8421L26.1456 34.2665H37.8542V41.0278C39.512 41.3765 41.0715 41.6169 42.5377 41.7643Z"
-        fill="#1C6DD0"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M11.0212 43.5317C15.6479 41.4569 23.2281 39.7504 35.6777 42.9932C43.296 44.9776 48.9864 44.7201 53.1723 43.5317C53.1723 43.7527 47.1883 56.2659 32.2066 56.2659C17.225 56.2659 11.0212 43.485 11.0212 43.5317Z"
-        fill="#1C6DD0"
-      />
-    </svg>
-  );
-};
-export default WebLogo;
